+++ Source/Model/Message/ZMClientMessage+Composite.swift
@@ -18,58 +18,18 @@
 
 import Foundation
 
-public protocol CompositeMessageData {
-    var items: [CompositeMessageItem] { get }
-}
+// MARK: - ConversationCompositeMessage
 
-public enum CompositeMessageItem {
-    case text(ZMTextMessageData)
-    case button(ButtonMessageData)
-    
-    internal init?(with protoItem: Composite.Item, message: ZMClientMessage) {
-        guard let content = protoItem.content else { return nil }
-        let itemContent = CompositeMessageItemContent(with: protoItem, message: message)
-        switch content {
-        case .button:
-            self = .button(itemContent)
-        case .text:
-            self = .text(itemContent)
+extension ZMClientMessage: ConversationCompositeMessage {
+    public var compositeMessageData: CompositeMessageData? {
+        guard case .some(.composite) = underlyingMessage?.content else {
+            return nil
         }
+        return self
     }
 }
 
-public protocol ButtonMessageData {
-    var title: String? { get }
-    var state: ButtonMessageState { get }
-    var isExpired: Bool { get }
-    func touchAction()
-}
-
-public enum ButtonMessageState {
-    case unselected
-    case selected
-    case confirmed
-    
-    init(from state: ButtonState.State?) {
-        guard let state = state else {
-            self = .unselected
-            return
-        }
-        self = ButtonMessageState(from: state)
-    }
-    
-    init(from state: ButtonState.State) {
-        switch state {
-        case .unselected:
-            self = .unselected
-        case .selected:
-            self = .selected
-        case .confirmed:
-            self = .confirmed
-        }
-    }
-}
-
+// MARK: - CompositeMessageData
 extension ZMClientMessage: CompositeMessageData {
     public var items: [CompositeMessageItem] {
         guard let message = underlyingMessage, case .some(.composite) = message.content else {
@@ -84,15 +44,7 @@ extension ZMClientMessage: CompositeMessageData {
     }
 }
 
-extension ZMClientMessage: ConversationCompositeMessage {
-    public var compositeMessageData: CompositeMessageData? {
-        guard case .some(.composite) = underlyingMessage?.content else {
-            return nil
-        }
-        return self
-    }
-}
-
+// MARK: - ButtonStates Interface
 extension ZMClientMessage {
     @objc static func updateButtonStates(withConfirmation confirmation: ZMButtonActionConfirmation,
                                          forConversation conversation: ZMConversation,
@@ -102,9 +54,20 @@ extension ZMClientMessage {
         message?.updateButtonStates(withConfirmation: confirmation)
     }
     
+    @objc static func expireButtonState(forButtonAction buttonAction: ZMButtonAction,
+                                        forConversation conversation: ZMConversation,
+                                        inContext moc: NSManagedObjectContext) {
+        let nonce = UUID(uuidString: buttonAction.referenceMessageId)
+        let message = ZMClientMessage.fetch(withNonce: nonce, for: conversation, in: moc)
+        message?.expireButtonState(withButtonAction: buttonAction)
+    }
+}
+
+// MARK: - ButtonStates Helpers
+extension ZMClientMessage {
     private func updateButtonStates(withConfirmation confirmation: ZMButtonActionConfirmation) {
         guard let moc = managedObjectContext else { return }
-
+        
         if !containsButtonState(withId: confirmation.buttonId) {
             ButtonState.insert(with: confirmation.buttonId, message: self, inContext: moc)
         }
@@ -115,14 +78,6 @@ extension ZMClientMessage {
         return buttonStates?.contains(where: { $0.remoteIdentifier == buttonId }) ?? false
     }
     
-    @objc static func expireButtonState(forButtonAction buttonAction: ZMButtonAction,
-                                        forConversation conversation: ZMConversation,
-                                        inContext moc: NSManagedObjectContext) {
-        let nonce = UUID(uuidString: buttonAction.referenceMessageId)
-        let message = ZMClientMessage.fetch(withNonce: nonce, for: conversation, in: moc)
-        message?.expireButtonState(withButtonAction: buttonAction)
-    }
-    
     private func expireButtonState(withButtonAction buttonAction: ZMButtonAction) {
         let state = buttonStates?.first(where: { $0.remoteIdentifier == buttonAction.buttonId })
         managedObjectContext?.performGroupedBlock { [managedObjectContext] in
@@ -130,115 +85,5 @@ extension ZMClientMessage {
             state?.state = .unselected
             managedObjectContext?.saveOrRollback()
         }
-    }
-}
-
-fileprivate class CompositeMessageItemContent: NSObject {
-    private let parentMessage: ZMClientMessage
-    private let item: Composite.Item
-    
-    private var text: Text? {
-        guard case .some(.text) = item.content else { return nil }
-        return item.text
-    }
-    
-    private var button: Button? {
-        guard case .some(.button) = item.content else { return nil }
-        return item.button
-    }
-    
-    init(with item: Composite.Item, message: ZMClientMessage) {
-        self.item = item
-        self.parentMessage = message
-    }
-}
-
-extension CompositeMessageItemContent: ZMTextMessageData {
-    var messageText: String? {
-        return text?.content.removingExtremeCombiningCharacters
-    }
-    
-    var linkPreview: LinkMetadata? {
-        return nil
-    }
-    
-    var mentions: [Mention] {
-        return Mention.mentions(from: text?.mentions, messageText: messageText, moc: parentMessage.managedObjectContext)
-    }
-    
-    var quote: ZMMessage? {
-        return nil
-    }
-    
-    var linkPreviewHasImage: Bool {
-        return false
-    }
-    
-    var linkPreviewImageCacheKey: String? {
-        return nil
-    }
-    
-    var isQuotingSelf: Bool {
-        return false
-    }
-    
-    var hasQuote: Bool {
-        return false
-    }
-    
-    func fetchLinkPreviewImageData(with queue: DispatchQueue, completionHandler: @escaping (Data?) -> Void) {
-        // no op
-    }
-    
-    func requestLinkPreviewImageDownload() {
-        // no op
-    }
-    
-    func editText(_ text: String, mentions: [Mention], fetchLinkPreview: Bool) {
-        // no op
-    }
-}
-
-extension CompositeMessageItemContent: ButtonMessageData {
-    var title: String? {
-        return button?.text
-    }
-    
-    var state: ButtonMessageState {
-        return ButtonMessageState(from: buttonState?.state)
-    }
-    
-    var isExpired: Bool {
-        return buttonState?.isExpired ?? false
-    }
-    
-    func touchAction() {
-        guard let moc = parentMessage.managedObjectContext,
-            let buttonId = button?.id,
-            let messageId = parentMessage.nonce,
-            !hasSelectedButton else { return }
-
-        moc.performGroupedBlock { [weak self] in
-            guard let `self` = self else { return }
-            let buttonState = self.buttonState ??
-                ButtonState.insert(with: buttonId, message: self.parentMessage, inContext: moc)
-            buttonState.state = .selected
-            self.parentMessage.buttonStates?.resetExpired()
-            self.parentMessage.conversation?.append(buttonActionWithId: buttonId, referenceMessageId: messageId)
-            moc.saveOrRollback()
-        }
-    }
-    
-    private var hasSelectedButton: Bool {
-        return parentMessage.buttonStates?.contains(where: {$0.state == .selected}) ?? false
-    }
-    
-    private var buttonState: ButtonState? {
-        guard let button = button else { return nil }
-
-        return parentMessage.buttonStates?.first(where: { buttonState in
-            guard let remoteIdentifier = buttonState.remoteIdentifier else { return false }
-            return remoteIdentifier == button.id
-        })
     }
 }
